prompt_engineering_lab/requirements.txt
torch
transformers
torchvision
pillow
matplotlib
Pillow<9.1
trimesh<3.10
pandas
pyglet<=1.5.0,>=1.4.0
gym>=0.9.0
numpy<1.20.0,>=1.10.0

pyzmq>=16.0.0
# scikit-image>=0.13.1
# opencv-python>=3.4
# opencv-python==4.3.0.36
opencv-python
#==4.4.0.44
pyyaml>=3.11

dataclasses<=0.6; python_version < '3.7'
carnivalmirror==0.6.2
pipdeptree

